For future submission

+++ push_swap/push_swap.h
@@ -6,7 +6,7 @@
 /*   By: knajmech <knajmech@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/12/19 08:37:36 by knajmech          #+#    #+#             */
-/*   Updated: 2026/02/06 07:27:11 by knajmech         ###   ########.fr       */
+/*   Updated: 2026/02/10 12:06:03 by knajmech         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -71,7 +71,7 @@ int		index_list_b(t_stack *stack_a, t_manager *heads, int start);
 void	spin_rev_b(t_stack *stack, t_stack *head);
 int		checker_b(t_stack *stack, t_manager *heads);
 int		valid_address(int compare, char s_or_l);
-int		ft_atoi_b(char *nbr);
+long	ft_atoi_b(char *nbr);
 void	execute_rule(t_manager *heads, int rule);
 int		assign_val(char *instruction, int move);
 void	pick_path_b(t_manager *heads, int num);
